Move the rules for requesting into a separate 'rules' namespace
